+++ src/strategyCalculator/OperationDiv.java
@@ -3,7 +3,7 @@ package strategyCalculator;
 public class OperationDiv implements Strategy {
 
 	@Override
-	public int DoOpertion(int a, int b) {
+	public float DoOpertion(float a, float b) {
 
 		if (b == 0) throw new IllegalArgumentException("El divisor es 0");
 		return a / b;
